+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.1.5 (2023-11-01)
+* Added ContentSecgurityPolicy Middleware
+* Added initial ErrorHandler implementation
+
 ## v0.1.4 (2023-11-01)
 * Removed time limit when sending responses
 * Added body HTML message to redirect response